requirements.txt
numpy==1.21.2
torch==1.11.0
torchvision==0.12.0
pytorch-lightning==1.6.0
timm==0.5.4
opencv-python-headless==4.5.5.64
albumentations==1.1.0
PyYAML==6.0
easydict

kafka-job-scheduler==1.0.*